install_k8s/gok-agent/v2/agent/requirements.txt
pika
python-jose
requests